Extract handler for shifted special keys
